+++ resources/acceleron-server/billing-core/CoreValidation.js
@@ -1,0 +1,16 @@
+"use strict";
+let BillingCoreUtils = require("./BillingCoreUtils");
+
+class CoreValidation {
+
+    constructor() {
+
+    }
+
+    //Check if all the menu items, are in sync with master menu
+    validateCartAgainstMasterMenu(masterMenu, orderCart) {
+        return orderCart;
+    }
+}
+
+module.exports = CoreValidation;
